+++ nrfutil/device/common.ts
@@ -5,21 +5,15 @@
  */
 
 import { getModule } from '../modules';
-import { Progress } from '../sandboxTypes';
+import { type OnProgress } from '../sandboxTypes';
 
 export const deviceTraitsToArgs = (traits: DeviceTraits) => {
-    const args: string[] = [];
     const traitsString = Object.keys(traits)
         .map(trait => (traits[trait as keyof DeviceTraits] ? trait : null))
         .filter(t => t !== null)
         .join(',');
 
-    if (traitsString.length > 0) {
-        args.push('--traits');
-        args.push(traitsString);
-    }
-
-    return args;
+    return traitsString.length > 0 ? ['--traits', traitsString] : [];
 };
 
 export type ResetKind =
@@ -157,7 +151,7 @@ export interface SerialPort {
 export const deviceSingleTaskEndOperation = async <T = void>(
     device: NrfutilDevice,
     command: string,
-    onProgress?: (progress: Progress) => void,
+    onProgress?: OnProgress,
     controller?: AbortController,
     args: string[] = []
 ) => {
@@ -178,7 +172,7 @@ export const deviceSingleTaskEndOperation = async <T = void>(
 export const deviceSingleTaskEndOperationVoid = async (
     device: NrfutilDevice,
     command: string,
-    onProgress?: (progress: Progress) => void,
+    onProgress?: OnProgress,
     controller?: AbortController,
     args: string[] = []
 ) => {
@@ -196,3 +190,5 @@ export const deviceSingleTaskEndOperationVoid = async (
         [...args, '--serial-number', device.serialNumber]
     );
 };
+
+export const coreArg = (core?: DeviceCore) => (core ? ['--core', core] : []);